requirements.txt
# Core
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# MCP
mcp>=1.0.0

# API & Frontend
fastapi>=0.115.0
uvicorn>=0.30.0
streamlit>=1.38.0

# Database
faker>=28.0.0

# Visualization
matplotlib>=3.9.0

# Logging & Monitoring
langsmith>=0.1.0

# Testing
pytest>=8.0.0
httpx>=0.27.0

# Deployment
huggingface-hub>=0.20.0
